+++ Frituquim/ViewModels/ExtractionViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
@@ -11,20 +12,21 @@ using Wpf.Ui.Controls;
 
 namespace Frituquim.ViewModels;
 
-public partial class ExtractionViewModel(ExtractionType extractionType, ISnackbarService snackbarService, ExecutionMessageMap messageMap) : FrituquimBasePageWithOutputDirectory
+public partial class ExtractionViewModel(
+    ExtractionType extractionType,
+    ISnackbarService snackbarService,
+    ExecutionMessageMap messageMap) : FrituquimBasePageWithOutputDirectory
 {
-    [ObservableProperty] private string _videoPathOrUrl = "https://www.youtube.com/watch?v=vaphaFCyLQI";
-    
+    [ObservableProperty]
+    private string _videoPathOrUrl = "https://www.youtube.com/watch?v=vaphaFCyLQI";
+
     [RelayCommand]
     private void OpenFileDialog()
     {
         var fileDialog = new OpenFileDialog();
-        if (fileDialog.ShowDialog() ?? false)
-        {
-            VideoPathOrUrl = fileDialog.FileName;
-        }
+        if (fileDialog.ShowDialog() ?? false) VideoPathOrUrl = fileDialog.FileName;
     }
-    
+
     [RelayCommand]
     private async Task Execute()
     {
@@ -35,16 +37,14 @@ public partial class ExtractionViewModel(ExtractionType extractionType, ISnackba
             var fileName = await YtdlpHelper.GetFileName(VideoPathOrUrl, ExtractionType.Audio);
             var downloadFilePath = Path.Combine(OutputDirectory, fileName);
 
-            if (File.Exists(downloadFilePath))
-            {
-                File.Delete(downloadFilePath);
-            }
+            if (File.Exists(downloadFilePath)) File.Delete(downloadFilePath);
 
-            var downloadedFile = await YtdlpHelper.CreateYtdlpCommand(VideoPathOrUrl, downloadFilePath, extractionType == ExtractionType.Audio ? ["-x", "--audio-format", "mp3"] : [])
+            var downloadedFile = await YtdlpHelper.CreateYtdlpCommand(VideoPathOrUrl, downloadFilePath,
+                    extractionType == ExtractionType.Audio ? ["-x", "--audio-format", "mp3"] : [])
                 .ExecuteAsync()
                 .Task
                 .ContinueWith(t => t.IsCompletedSuccessfully);
-            
+
             if (!downloadedFile)
             {
                 snackbarService.Show(messageMap.Error.Title, messageMap.Error.Message, ControlAppearance.Danger,
@@ -56,10 +56,7 @@ public partial class ExtractionViewModel(ExtractionType extractionType, ISnackba
             snackbarService.Show(messageMap.Success.Title, messageMap.Success.Message, ControlAppearance.Success,
                 null, TimeSpan.FromSeconds(3));
 
-            if (OpenFolderAfterExecution)
-            {
-                System.Diagnostics.Process.Start("explorer", OutputDirectory);
-            }
+            if (OpenFolderAfterExecution) Process.Start("explorer", OutputDirectory);
         }
         finally
         {
